requirements.txt
boto3==1.24.17
Django==4.0.5
pyngrok==5.1.0
python-decouple==3.6
twilio==7.9.3
node==0.9
requests~=2.28.0